+++ flutter_plugin_qpos/CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.1.0+1
+## 0.1.1
 
-* 1. update Gradle version and compile version
+* 1. Add mifare card operation on the ios
+* 2. Update Bluetooth permission for android 12
+* 3. Update SDK version to 4.0.0